// supabase/functions/_shared/generated/kosha-catalog.ts
export interface ConstructionProcessCatalogItem {
  projectType: string;
  tradeName: string;
  detailProcess: string;
  tokens: string[];
}

export interface MachineEquipmentCatalogItem {
  majorIndustry: string;
  middleIndustry: string;
  subIndustry: string;
  equipmentName: string;
  equipmentNameEn: string;
  tokens: string[];
}